🙉 Improve bananalyzer
