+++ bananalyzer/runner/runner.py
@@ -1,13 +1,15 @@
 # dynamic_test_runner.py
 import os
 import tempfile
-from typing import IO, Awaitable, Callable, List
+from typing import IO, Awaitable, Callable, Dict, List
+from urllib.parse import urlparse
 
 import pytest
 from pydantic import BaseModel
 
+from bananalyzer import AgentRunner
 from bananalyzer.data.schemas import Example
-from bananalyzer.schema import PytestArgs
+from bananalyzer.schema import AgentRunnerClass, PytestArgs
 
 TestType = Callable[[], Awaitable[None]]
 
@@ -17,48 +19,66 @@ class BananalyzerTest(BaseModel):
     example: Example
 
 
-def generate_test(example: Example, headless: bool) -> BananalyzerTest:
-    return BananalyzerTest(
-        code=f"""
+class TestGenerator:
+    def __init__(self) -> None:
+        self._names: Dict[str, int] = {}
+
+    def generate_test(self, example: Example, headless: bool) -> BananalyzerTest:
+        return BananalyzerTest(
+            code=f"""
 @pytest.mark.asyncio
-async def test_{example.id.replace("-", "_")}() -> None:
+async def {self._generate_name(example)}() -> None:
     async with async_playwright() as p:
         browser = await p.chromium.launch(headless={headless})
         context = await browser.new_context()
         example = get_example_by_url("{example.url}")
-         
-        # The agent is imported into the global context prior to this call 
+
+        # The agent is imported into the global context prior to this call
         result = await agent.run(context, example)
         for curr_eval in example.evals:
             curr_eval.eval_results(p, result)
-""",
-        example=example,
-    )
+    """,
+            example=example,
+        )
+
+    def _generate_name(self, example: Example) -> str:
+        domain = urlparse(example.url).netloc
+        domain = domain.replace(".", "_")
+        if domain.startswith("www_"):
+            domain = domain[4:]
+
+        key = f"{example.type}_{domain}"
+        self._names[key] = self._names.get(key, -1) + 1
+        suffix = "" if not self._names[key] else f"_{self._names[key]}"
+        return f"test_{key}{suffix}"
 
 
-def write_load_agent_module(f: IO[str], file_path: str) -> None:
+def write_load_agent_module(f: IO[str], runner: AgentRunnerClass) -> None:
     f.write(
         f"""
 import importlib.util
 import sys
 from pathlib import Path
+import pprint
 
 def load_agent_module():
-    path = Path("{file_path}")
+    path = Path("{runner.class_path}")
     module_name = path.stem
 
-    spec = importlib.util.spec_from_file_location(module_name, "{file_path}")
+    spec = importlib.util.spec_from_file_location(module_name, str(path))
     module = importlib.util.module_from_spec(spec)
     sys.modules[module_name] = module
     spec.loader.exec_module(module)
-    globals()["agent"] = getattr(module, "agent")
+    print(f"Loaded agent module", module)
+    globals()["agent"] = getattr(module, '{runner.class_name}')()
+    print(f"Loaded agent", agent)
 load_agent_module()
 """
     )
 
 
 def create_test_file(
-    tests: List[BananalyzerTest], prefix: str, agent_file_path: str
+    tests: List[BananalyzerTest], prefix: str, runner: AgentRunnerClass
 ) -> str:
     with tempfile.NamedTemporaryFile(
         mode="w+", delete=False, prefix=prefix, suffix=".py"
@@ -68,7 +88,7 @@ def create_test_file(
         f.write("from playwright.async_api import async_playwright\n\n")
         f.write("global test\n")
         f.write("test = 'test'\n")
-        write_load_agent_module(f, agent_file_path)
+        write_load_agent_module(f, runner)
 
         for test_content in tests:
             f.write(f"{test_content.code}\n\n")
@@ -77,7 +97,7 @@ def create_test_file(
 
 
 def run_tests(
-    tests: List[BananalyzerTest], agent_file_path: str, pytest_args: PytestArgs
+    tests: List[BananalyzerTest], runner: AgentRunnerClass, pytest_args: PytestArgs
 ) -> int:
     """
     Create temporary test files based on intent, run them, and then delete them
@@ -87,7 +107,7 @@ def run_tests(
         [test for test in tests if test.example.type == intent] for intent in intents
     ]
     test_file_names = [
-        create_test_file(tests, f"{tests[0].example.type}_intent_", agent_file_path)
+        create_test_file(tests, f"{tests[0].example.type}_intent_", runner)
         for tests in intent_separated_tests
     ]
 
@@ -96,6 +116,7 @@ def run_tests(
             test_file_names
             + (["-s"] if pytest_args.s else [])
             + ([f"-n {pytest_args.n}"] if pytest_args.n else [])
+            + ["-v"]
         )
         return_code = pytest.main(args)
     finally:
