--- bananalyzer/test_bananalyzer_agent.py
@@ -1,6 +1,0 @@
-"""
-Test bananalyzer file to ensure CLI is working as expected within the project
-"""
-from bananalyzer.runner.null_agent_wrapper import NullAgentRunner
-
-agent = NullAgentRunner()
