+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "bananalyzer"
-version = "0.3.5"
+version = "0.4.0"
 description = "Open source AI Agent evaluation framework for web tasks 🐒🍌"
 authors = ["asim-shrestha <asim.shrestha@hotmail.com>"]
 readme = "README.md"
@@ -51,7 +51,7 @@ warn_unused_ignores = false
 warn_return_any = false
 namespace_packages = true
 files = "bananalyzer"
-exclude = ["tests"]
+exclude = ["tests", "venv"]
 
 [build-system]
 requires = ["poetry-core"]
